# openapi/buildsafe/v1/app.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "buildsafe/v1/app.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AppService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/apps": {
      "get": {
        "operationId": "AppService_ListApps",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAppsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "AppService"
        ]
      },
      "post": {
        "operationId": "AppService_RegisterApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RegisterAppResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RegisterAppRequest"
            }
          }
        ],
        "tags": [
          "AppService"
        ]
      }
    },
    "/v1/apps/{app_id}": {
      "get": {
        "operationId": "AppService_GetApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAppResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "AppService"
        ]
      },
      "delete": {
        "operationId": "AppService_UnregisterApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UnregisterAppResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "AppService"
        ]
      }
    },
    "/v1/apps/{app_id}/details": {
      "post": {
        "operationId": "AppService_AddAppDetails",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddAppDetailsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppServiceAddAppDetailsBody"
            }
          }
        ],
        "tags": [
          "AppService"
        ]
      }
    }
  },
  "definitions": {
    "AppServiceAddAppDetailsBody": {
      "type": "object",
      "properties": {
        "branch": {
          "type": "string"
        },
        "folder": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddAppDetailsResponse": {
      "type": "object"
    },
    "v1App": {
      "type": "object",
      "properties": {
        "app_id": {
          "type": "string",
          "format": "uint64"
        },
        "project_id": {
          "type": "string",
          "format": "uint64"
        },
        "repo_name": {
          "type": "string"
        },
        "repo_id": {
          "type": "string"
        },
        "branch": {
          "type": "string"
        },
        "folder": {
          "type": "string"
        }
      }
    },
    "v1GetAppResponse": {
      "type": "object",
      "properties": {
        "app": {
          "$ref": "#/definitions/v1App"
        }
      }
    },
    "v1ListAppsResponse": {
      "type": "object",
      "properties": {
        "apps": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1App"
          }
        }
      }
    },
    "v1RegisterAppRequest": {
      "type": "object",
      "properties": {
        "project_id": {
          "type": "string",
          "format": "uint64"
        },
        "repo_name": {
          "type": "string"
        },
        "repo_id": {
          "type": "string"
        }
      }
    },
    "v1RegisterAppResponse": {
      "type": "object",
      "properties": {
        "app_id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1UnregisterAppResponse": {
      "type": "object"
    }
  }
}